Move and rename after merge
